Added detecting ext2fs (#10)
* Added detecting ext2fs

* Correctetec multiple block read and csd read

* Commented out PWR pin

* Corrected identifying sd version

+++ src/main/fs/ext2/inode_descriptor.h
@@ -1,9 +1,12 @@
 #ifndef MEXT2_FS_EXT2_H
 #define MEXT2_FS_EXT2_H
 
+#include <stdint.h>
+#include "common.h"
+
 struct mext2_inode_descriptor
 {
-    uint64_t inode_pos;
+    uint32_t inode_pos;
     uint32_t i_size;
     uint32_t i_blocks;
     uint32_t i_dir_acl;
